refactor(GameResult): GameResult 클래스명 MatchResult로 변경
